# public/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema Bibi - ServiceOS — API
  version: 3.0.8
  description: >
    API do Sistema Bibi - ServiceOS v3.0.7 — plataforma Pay Per Use multi-nicho. Multi-tenant,
    quatro portais segregados por `role`, vocabulário adaptável por segmento.

    Autenticação por cookie de sessão httpOnly (`bibi_session`) assinado via
    HMAC-SHA256, obtido em `POST /api/auth/login`.

    Fluxos: docs/produto/FLUXOS.md · Arquitetura: docs/plataforma/ARQUITETURA.md
...
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
  - url: https://sistema-bibi.netlify.app
    description: Produção (Netlify)
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bibi_session
  schemas:
...
    Usage:
      type: object
      properties:
        id: { type: string }
        procedure: { type: string }
        category: { type: string, enum: [CONSULTA, EXAME] }
        priceCharged: { type: number }
        priceLabel: { type: string, example: "R$ 220,00" }
        billed: { type: boolean }
    AppointmentDetail:
      type: object
      properties:
        appointment:
          type: object
          properties:
            id: { type: string }
            scheduledAt: { type: string, format: date-time }
            status: { type: string }
            reason: { type: string, nullable: true }
        patient:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            cpf: { type: string }
            company: { type: string, nullable: true }
            companyId: { type: string, nullable: true }
        usages:
          type: array
          items: { $ref: '#/components/schemas/Usage' }
        records:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              content: { type: string }
              createdAt: { type: string, format: date-time }
...
security:
  - sessionCookie: []
paths:
...
  /api/calendar/oauth/{provider}/start:
    parameters:
      - name: provider
        in: path
        required: true
        schema: { type: string, enum: [google, microsoft] }
      - name: scope
        in: query
        schema: { type: string, enum: [PROVIDER, TENANT] }
      - name: returnTo
        in: query
        schema: { type: string }
    get:
      tags: [Público]
      summary: Inicia OAuth Google/Microsoft Calendar
      responses:
        '302': { description: Redirect ao provedor }
        '503': { description: Credenciais OAuth não configuradas }
...
  /api/prestador/appointments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Prestador]
      summary: Detalhe do atendimento (procedimentos + prontuário)
      responses:
        '200':
          description: Detalhe do atendimento
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppointmentDetail' }
        '404': { description: Agendamento não encontrado }
    patch:
      tags: [Prestador]
      summary: Atualiza o status do atendimento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [AGENDADO, CONFIRMADO, REALIZADO, FALTOU, CANCELADO]
      responses:
        '200':
          description: Status atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  status: { type: string }
        '400': { description: Status inválido }
        '404': { description: Agendamento não encontrado }
...